destroyed functions folder and separated it
