+++ src/robot/index.ts
@@ -1,3 +1,11 @@
 // Main exports from the robot module
 export { RobotBuilder } from './robot-builder';
-export type { RobotComponents, RobotConfig } from './types';
+export type { RobotComponents } from './types';
+export {
+    createLinkOrigin,
+    createLink,
+    createLinkEndFrame,
+    createJoint,
+    createJointFrame,
+    createTCPframe,
+} from './functions';
