+++ src/scene/index.ts
@@ -2,3 +2,11 @@
 export { SceneManager } from './scene-manager';
 export { defaultSceneConfig } from './config';
 export type { SceneConfig, SceneComponents } from './types';
+export {
+    createScene,
+    createRenderer,
+    createCamera,
+    createGrid,
+    createControls,
+    createReferenceFrame,
+} from './functions';
