+++ src/shared/index.ts
@@ -1,0 +1,2 @@
+// Shared types and utilities
+export type { Axis, CylinderMesh } from './types';
